[IMP] mass_mailing: allow users to re-use old mass mailing template
Purpose
=======
Allow users to re-use a design they made in a previous mailing.

Specifications
==============
Add a duplicate button when the mailing has been sent (so the users
can easily duplicate an existing mailing to re-use the template).

When we open the web editor for the first time, show the email bodies
of the 8 last mailings sent. When the users click on them, load the
HTML like we do for the themes.

Technical
=========
Preview: we need to display a preview of the template. Existing themes
have a screenshot of the theme but for our case it's not possible
(it would require to generate the image and store it in the server side
for each mailing...). So instead we use an iframe. The content of the
iframe will be the `body_html` and not `body_arch`. This is because
`body_html` is basically `body_arch` where we replaced all bootstrap
CSS classes into inline CSS. As the iframe do not share the CSS of it's
parent, this is needed to correctly make the preview (otherwise content
might do not look like it should).

When the user click on the template, we set the current body to
`body_arch` (and not `body_html`) to keep the content editable.

Task-2593231