[FIX] microsoft_calendar: multi-calendar sync support + bug fixes
The main goal of this commit is to fix several issues which happen
when several Odoo users want to sync their Odoo calendar with their
Outlook calendar and when they are organizer/attendees of a same event.

The actual version of `microsoft_calendar` only use the `id` field of
an Outlook event to map it with an Odoo event. This is not enough as this
`id` is specific to a Outlook user calendar. That means, a same event will
not have the same `id` in the organizer Outlook calendar than in the
attendee Outlook calendar.

To solve that, we have to use the `iCalUId` field which identifies an event
among all Outlook calendars where it is present. Unfortunately, as the Outlook API
has been designed to be used with the event `id` and not with its `iCalUId`,
we have to juggle with both the organiser calendar event `id` and the `iCalUid`.

This commit is also an opportunity to clean the code a little bit and to add some
comment to better understand what some methods do.

opw-2593661