[WIP] l10n_fr: clean tax groups, split taxes TVA IC
This commit cleans the tax groups:
- it removes the negative tax groups
- it makes sure that there is only one tax group per tax rate
- all the taxes with the same rates are attached to a single tax group

This commit also splits the taxes "TVA Intracommunautaire (achat) X%" into:
- "TVA IC (achat) X% Biens"
- "TVA IC (achat) X% Services"
(with X taking the values 20, 10, 8.5, 5.5, 2.1)

This commit creates the tax "TVA IC 0% Services" as in Didier Six's email. The tax "TVA IC 0% Biens" should then also be created ? With which tax grids and accounts ?

This commit creates the tax "TVA extra UE 20% Services" as in Didier Six's email. The tax "TVA extra UE 20% Biens" should then also be created ? Which tax grids and which accounts ? Should also be created for the TVA rates 10, 8.5, 5.5, 2.1 ?

For the moment "TVA IC 0% Biens" and "TVA extra UE 20% Biens" were added but the grids and accounts are WRONG ! They need to be modified.

task-2674047