http.py json_dispatch and a bit of RPC
The implementation of `@route(type='json')`. If you were expecting a
simple implementation such as

    self.params = json.loads(self.httprequest.get_data())
    result = endpoint(**self.params)
    return Response(json.dumps(result), 200, [
        ('Content-Type', 'application/json),
    ]

you might be surprise to read about JSON-RPC2. Yup some guy decided it
would be cute to add a bunch of useless things (all three jsonrpc,
method and id are basically ignored). Also that context stuff? Just
throw it in the params, nobody will notice (I did :eyes:). The error
handling has some gotchas too (but it is coming in another commit).

The best is that most endpoints type='json' are... implementing another
RPC layer on top of JSON-RPC2. Typical request:

    POST /web/dataset/call_kw HTTP/1.1
    Content-Type: application/json

    {
        jsonrpc: "2.0",
        id: null,
        method: 'call',
        params: {
            'model': 'res.partner',
            'method': 'create',
            'args': [...],
            'kwargs': {},
            'context': {},
        }
    }

Honorable mentions: /jsonrpc, /web/dataset/*

Why am I so salty? It is because there is nothing I can do to change the
current implementation. Changing the implementation would break the
current API which is working fine. It would be a cross-team effort as
both the JS Framework and the SaaS would be impacted. By the way, there
was an effort to reduce the cumbersome of the XML-RPC layer, they are
odoo/odoo#3989 and odoo/odoo#67388, #3989 is 7 years old... yeah... not
doing it again for jsonrpc. I blame nobody, JSON-RPC is cute indeed,
just... not very useful here :( (same goes for XML-RPC without #3989)

The `dispatch_rpc` fonction have been moved to the `service` module
where it really belongs. The two options --log-request and
--log-response were pouring the logs with useless informations,
sometimes twice (i.e. /jsonrpc). The logging level was debug because it
annoys people to have the whole `params` logged. It would be better to
introduce a bit of logging (level info) in every "rpc over rpc"-like
controller methods. Duty for another PR? I myself use mitmproxy whenever
I want the complete logs (request+response) of the web server.

By the way, the logs in `service/server.py` were not using the "correct"
logger name, it seemed dead code to me so I removed it too, heh.