[FIX] website: prepare cover dataset upon color selection
Before this commit the cover dataset was only updated during updateUI()
but if the cover was saved before closing the color palette after
selecting a solid color, the selected color was not confirmed and the
saved cover dataset was based on the previously selected color.

After this commit the cover dataset is updated whenever an option is
selected to handle the specific "save before close color palette" case.
Because this is not done after updateUI anymore, the other options
values are obtained from the DOM instead of from the options themselves.

task-2678100