[FIX] mail: send mails with too heavy attachments
Emails that are too heavy to be supported by the mail server
are stripped of their attachments. Instead of sending the attachment(s)
through mail, we add a note that informs the recievers that
the attachment(s) are available on portal.

Emails that were too heavy weren't send at all. Now we notify the
recievers of the chatter update, even when we cannot send the
attachment(s).

task-2474726