[FIX] mail, various: remove unused rendering option in QWeb
Render safe option was used in jinja rendering to tell jinja that content
was safe and was not expected to be escaped. This was used notably in subject
as characters like ``&`` was escaped in email subject.

This is now supported directly by the "inline template" rendering engine that
is used to render text-based templates. This option is not supported by code
anymore anyway.

Task-2673913 (TestMail: cleanup posting/composer tests)
Task-2088884 (MailComposer: onchange to computed stored fields)