[WIP] web_editor: add image shapes with perspective 3D
This commit is a work in progress to add image shapes with a 3D
perspective. The purpose is to use a foreignObject tag to be able to
wrap the image into a div and use the perspective attribute on it.
The current issue is the foreignObject tag doesn't scale properly on
Safari which gives bad render most of the time.