[IMP] web_editor: permit the user to set link on images
Allow the user to set a link on an image directly via a new tool in the
web editor. The user can access this link tool in the editor or via the
CTRL+K shortcut. This tool is very similar to the existing link tool
for texts. This tool uses the suggestion system already existing on the
text link tool. Some utility functions had to be migrated from website
to web_editor.

task-2666467