[IMP] sale_project: add a new field in product template
In this commit, we add a 'Tasks Templates' m2m field below the
'project template' field. The field is only visible if 'project & task'
is selected for 'create on order' and if the 'project template' field
is set on a product. Also, if  'Task Template' is false then it copies
all of the tasks from the selected project template on the SO
confirmation else, only copy the selected tasks.

task-2672783