[IMP] base, mail, *: add value labels for boolean fields
Purpose
=======
"True" and "False" are not very user-friendly, especially for
non-technical people. Instead, we want to have a label for each value
(e.g. "Inactive" instead of "False" for the active field).

Specification
=============
This change is done for the import / export, the custom filter and
the tracking messages.

For the import / export, the label is exported in the current language
and when we import it again, we check for those translations. When no
labels are defined on the field, we use "Yes" and "No" instead of "1"
and "0" so the generated file is more clear for non-technical people.

The tracking message is similar to the import / export feature.

For the custom filter, the labels are hardcoded in a mixin and can
therefor not be changed without refactoring this part of the code. So
for now, we use "Yes" and "No" instead.

Task-2648390