[FIX] stock_account:make cogs multi_company safe
The aim of this commit is to make the cogs generation multi-company
safe.

Before this commit:
When posting an invoice from company B with both company A and B
selected and A being selected as "main", the cost from company A could
be selected instead of cost from company A to create the anglosaxon
lines.

After this commit:
Whatever the selected companies are, the cost selected will always be from
the company that created the invoice.

task: #2686104