[IMP] lunch: add 'Sent' status and allow ordering in the future
Added a new 'Sent' status, meaning the orders have been made to the
supplier. Once an order is 'Sent' it cannot be modified anymore.

It's no longer possible to order a product after the automatic email has
been sent, if the supplier is available the next day, the customer can
add it to their basket for that day instead.

TaskID: 2678064