[l10n_ch] account: allow for QR bills to be printed in batch
The only option to print a QR Invoice was to go on the invoice page and to click on the Print QR Button.

This option is now available amongst the Print options when several Invoices are selected.

If one of the Invoices can't be printed under the QR format, its number is displayed with the error message.