[IMP] website: provide clarity for theme colors
Prior to this commit, the color presets generated by the color palette
would always be visible by the user which could lead to confusion.
Most users would click on the presets and edit them instead of editing
the color palette

This commit groups the presets into a collapsed section under the color
palette and changes the cursor into a pointer to intend that the color
palette is editable.

taks-2687469