[IMP] website: add content on hover
Add an option in the editor to activate custom content when hovering
specific elements. This option adds an invisible element on top of the
original element that will only be shown to the visitor when hovering
the original element.

A second option in the editor allows the user to edit the content shown
on hover. This option, when enabled, forces the hover content to be
shown even when not hovering, as to allow the end user to freely make
changes. When disabled, the opposite happens, the hover content will
stay hidden, so the original content can be edited.

When exiting the snippet editor while the edit option is enabled, this
behaviour is kept, so that other snippets can be added to the custom
hover content.

task-2666891

Co-Authored-By: Younn Olivier <yol@odoo.com>