[FIX] website: fix navbar menu label edition
When changing the label of a navbar menu via the link popover and the
menu edition dialog, the label would not be changed after saving.
Reloading the page would display the updated label.

While the RPC was correctly made to update the website.menu, the editor
would apply a historyRevert after the DOM update to change the menu
label on the page. This change should not be recorded by the editor as a
RPC to update the website.menu is executed at the same time.

task-2513588