[IMP] event: avoid concurrent_update errors for seats fields
When an email is sent to many people about an event with a registration link,
many users may want to register at the same time. In that case, the database
can be flooded with update queries for the stored values for the event(ticket)
seat availability (such as reserved, expected, and available).

In this commit, we remove the `store=True` attribute of those fields, and
therefore remove the Reporting/Event feature depending on them and rewrite
some domain searches and _compute fields in the event and event_sale modules.

Task-2654816