[FIX] web_editor: fix document display after using media dialog
The documents uploaded via the media selector were using an empty link
with the class o_image, displaying the icon relative to the document
type. There was an issue with the editor, filling the content of such
links with the source.

Now, when a document is selected via the media dialog, a link with the
data attribute fileName and content set as the filename.
The content of the links with o_image class is not filled.

The user can then use the new documents options in the link tools to
change the document associated to the created link, or the opening
method (download/open in a new page).

task-2172311