[IMP] mrp: enable custom consumption on components
Currently, when changing qty_producing on a MO, consumed number of each
component will always be updated. In this commit, we add a new boolean
field custom_consumption to each line. If checked, that line won't be
updated when changing qty_producing.
Same custom_consumption field is also added to mrp.bom.line as the
default value used when creating a MO.

Task-2687494