[IMP] website: give more flexibility to dynamic snippet templates
wip message :commit [1] introduced new templates. this commit gives
more flexibility to the template system.
adds a data-row-per-slide to dynamic snippet
templates so that you can have more than one row in templates.
Adds a data attributes to a template that allows to move
the position of the arrows.
Also add a data-extra-classes so that a template can change
the style of the overall snippet rather than just the card.

[1]: 49bd79b0bdca76415780bdfa199195371c2692e5

task-2677203