[FIX] mrp: Template of variant no longer show forecasted stock
Current Behaviour :
When having a product template with multiple variant, the forecasted stock and on_hand quantity are incorrect when the variant have Kit BoM. This lead to information that might be misinterpreted.

Behaviour after the PR :
To solve this issue, stock and forecasted stock are now hidden from the product template with multiple variants as they and should be seen on product.product.

opw-2678400