[FIX] project: fix default assigned to for rating in project
Purpose of this this commit to add default assigned to for rating
when rating given for specific task and task have only one
assignees.

So, In this commit overwrite rating_get_rated_partner_id method
and set rated_partner_id depending on user set on task and if
task contains more then one user then keep it false.

task-2696911
closes: #80232