[IMP] web_editor: allow shapes on gifs
To apply shapes on images, we need to store them
(in base64) inside the shape SVG, but before that,
we need to use 'applyModifications()' to get the
dataURL that contains the image with applied filters
and size options...(see ImageOptimize > _writeShape).
But since we can't handle GIFs on applyModifications(),
we get a single frame of the animation as a result...

The goal of this commit is to skip the updates from
'applyModification()' (when the target is a .GIF) and
only apply the shape.

This way, other "data-attributes" options should be
disabled for GIFs and only shapes option is allowed.

task-2679905