[IMP] website_sale: change card background in dynamic snippet
With commit [1], a few templates were introduced that use bootstrap's
card classes. While most of them work fine, "Borderless 1 & 2"
had price made unreadable in certain color presets.

For this reason, it was decided to change the template to instead
have an "opacity" background : plain white background with an opacity of
.2 to make the card stand out but still have a color template that's
close to what the user chose.

[1]: 49bd79b0bdca76415780bdfa199195371c2692e5
task-2677203