anv review, allow Content-Type:json in http routes
Some payment providers are sending requests on our webhooks to confirm a
payment. Often they send the information json-serialized in the body. We
cannot declare the route as type='json' because they don't implement
json-rpc2 (normal, their requests are no RPC), we cannot declare the
route as type='http' because of the silly verification "the content type
of the request must match the type of the route". We drop that
verification. type=json is broken QED.