[FIX] website: cookies modal not scrollable
Steps to reproduce:
- Activate the cookie dialog from the website settings
- Remove the backdrop from the cookie dialog
- Add some blocks to the cookie dialog in the website
editor until the dialog height is bigger than the page
height.

Result:
The cookie dialog is not scrollable, so it's impossible
to reach the "I agree" button and close the dialog,
leaving the user stuck at the first page.

I check if the popup content is higher than the window.
If it is, I give priority to the popup scroll over the page one.
I had to overwrite the _showElement method to call
_setScrollbar after the content rendering,
in this way I can get its correct height.

Also I have to set 'pointer-events' to none
only if the popup is smaller than the page itself
(to interact with and scroll the page behind),
otherwise the popup won't be scrollable.
I did this by adding a class "s_popup_overflow_page"
if the popup is higher than the page and I adapted
the pointer-events property accordingly.

opw-2660786